refactor: (#60) MealMenu e2e 도메인 단위 통합
